+++ models/__init__.py
@@ -1,0 +1,4 @@
+from models.car_model import Car
+from models.user_model import User
+
+__all__ = ['Car', 'User']
